+++ changelog.markdown
@@ -5,7 +5,7 @@
 + Now attribute 'iconRange' is represented as a list of object with option to choose starOn and starOff;
 + Added full round value without half star option is enabled;
 + Added attribute 'showHalf' to just display and separate from the option to vote 'half';
-+ Added attribute 'targetOutValue' to be applied on 'target' when the 'targetKeep' is disabled;
++ Added attribute 'targetOutValue' to choose a default value to be setted into target on mouseout.
 + Added attribute 'precision' to be able rating with precision values, without star representation;
 + Added attribute 'space' to be able take off the spaces between the star;
 + Fixed bug in IE 7 that returns void instead undefined when there is no an attribute;